pyproject.toml
[project]
name = "grassmann"
version = "0.1.0"
description = "Attempt at reproducing 'Attention Is Not What You Need: Grassmann Flows as an Attention-Free Alternative for Sequence Modeling.' (Zhang, C. 2025)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "hydra-core>=1.3.2",
  "pyarrow>=18.1.0",
  "datasets>=3.2.0",
  "matplotlib>=3.9.0",
  "torch>=2.10.0",
  "tqdm>=4.67.0",
  "transformers>=4.48.3",
]
